requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.15.0
pytesseract>=0.3.8
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
Pillow>=8.3.0
matplotlib>=3.3.0
scipy>=1.7.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5